+++ changelog.md
@@ -6,6 +6,7 @@
 > This version is not released yet and is under active development.
 
 - Allow union, intersection, difference and symmetric difference of `Group`.
+- Implements `<=` , `<`, `>=`, `>`, `|`, `&`, `-` and `^`  operators for `Group`.
 - Deduplicate platforms on `Group` instantiation.
 - Allow testing for membership of individual platform in `Group`.