+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.5.1 (unreleased)](https://github.com/kdeldycke/extra-platforms/compare/v1.5.0...main)
+
+> [!IMPORTANT]
+> This version is not released yet and is under active development.
+
 ## [1.5.0 (2024-11-10)](https://github.com/kdeldycke/extra-platforms/compare/v1.4.0...v1.5.0)
 
 - Allow union, intersection, difference and symmetric difference of `Group`.